Order and order detail fully implemented
OrderController implements update an order.
OrderDetailController implements create, update (quantity and status) and delete an order detail.
OrderDetailRequest includes the id of order detail and the order id.
OrderRequest removes the date (it's included in the transaction), and includes the actual delivery date.
OrderResponse removes the date.
OrderDetailMapper trying to map a null status (not included in request) will return a null. This is required for some requests.
OrderMapper trying to map a null status and details (not included in request) will return a null. This is required for some requests.
Order model removes the date.
OrderDetailStatus removes the canceled status that is not needed.
OrderRepository includes functions to search by provider code.
OrderDetailService implements methods to update quantity, update status, delete an order detail, create a new order detail into a order and recalculate how much has been paid from an order.
OrderService fixes a problem where the paid, owed and total were not set appropriately. Now it has a method to update a order (change the dates, status and notes).
Exceptions added.