Order endpoint updated
Controller: Adds a method to deliver all the details in a order, to update the status of an order, and adds the new optional parameters (dates) for the search method.
Mapper: Fixed a validation in the get order method if the status is empty.
Repository: Adds methods to search orders by provider code, dates, actual and expected delivery dates.
Service: Adds a method to change the state of every order detail in a order to delivered and the search method now searches orders by date, delivery dates and provider code.